Move Bridge and GliX bridge into dedicated module.
