+++ sieves/tasks/predictive/question_answering/core.py
@@ -10,7 +10,8 @@ from sieves.data import Doc
 from sieves.engines import Engine, EngineType, dspy_, glix_, instructor_, langchain_, ollama_, outlines_
 from sieves.engines.core import EngineInferenceMode, EngineModel, EnginePromptSignature, EngineResult
 from sieves.serialization import Config
-from sieves.tasks.predictive.core import GliXBridge, PredictiveTask
+from sieves.tasks.predictive.bridges import GliXBridge
+from sieves.tasks.predictive.core import PredictiveTask
 from sieves.tasks.predictive.question_answering.bridges import (
     DSPyQA,
     InstructorQA,
