+++ sieves/tasks/predictive/summarization/bridges.py
@@ -9,7 +9,7 @@ import pydantic
 
 from sieves.data import Doc
 from sieves.engines import EngineInferenceMode, dspy_, instructor_, langchain_, ollama_, outlines_
-from sieves.tasks.predictive.core import Bridge
+from sieves.tasks.predictive.bridges import Bridge
 
 _BridgePromptSignature = TypeVar("_BridgePromptSignature")
 _BridgeResult = TypeVar("_BridgeResult")
@@ -24,20 +24,20 @@ class SummarizationBridge(
         task_id: str,
         prompt_template: str | None,
         prompt_signature_desc: str | None,
-        max_n: int,
+        n_words: int,
     ):
         """
         Initializes InformationExtractionBridge.
         :param task_id: Task ID.
         :param prompt_template: Custom prompt template.
         :param prompt_signature_desc: Custom prompt signature description.
-        :param max_n: Maximal number of words (consider this a guideline, not a strict limit).
+        :param n_words: Approximate number of words in summary.
         """
         super().__init__(task_id=task_id, prompt_template=prompt_template, prompt_signature_desc=prompt_signature_desc)
-        self._max_n = max_n
+        self._n_words = n_words
 
     def extract(self, docs: Iterable[Doc]) -> Iterable[dict[str, Any]]:
-        return ({"text": doc.text if doc.text else None, "max_n": self._max_n} for doc in docs)
+        return ({"text": doc.text if doc.text else None, "n_words": self._n_words} for doc in docs)
 
 
 class DSPySummarization(SummarizationBridge[dspy_.PromptSignature, dspy_.Result, dspy_.InferenceMode]):
